Track read bytes with index as READ_BYTES
