+++ test/unit/org/apache/cassandra/sensors/SensorsReadTest.java
@@ -17,6 +17,8 @@
  */
 package org.apache.cassandra.sensors;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 import org.junit.After;
@@ -26,7 +28,9 @@ import org.junit.Test;
 
 import org.apache.cassandra.SchemaLoader;
 import org.apache.cassandra.Util;
+import org.apache.cassandra.cql3.ColumnIdentifier;
 import org.apache.cassandra.cql3.Operator;
+import org.apache.cassandra.cql3.statements.schema.IndexTarget;
 import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.DecoratedKey;
 import org.apache.cassandra.db.Keyspace;
@@ -35,11 +39,16 @@ import org.apache.cassandra.db.ReadCommandVerbHandler;
 import org.apache.cassandra.db.RowUpdateBuilder;
 import org.apache.cassandra.db.compaction.CompactionManager;
 import org.apache.cassandra.db.marshal.AsciiType;
+import org.apache.cassandra.db.marshal.LongType;
+import org.apache.cassandra.dht.Murmur3Partitioner;
+import org.apache.cassandra.index.sai.StorageAttachedIndex;
 import org.apache.cassandra.io.sstable.format.SSTableReader;
 import org.apache.cassandra.io.util.DataOutputBuffer;
 import org.apache.cassandra.net.Message;
 import org.apache.cassandra.net.MessagingService;
 import org.apache.cassandra.net.Verb;
+import org.apache.cassandra.schema.IndexMetadata;
+import org.apache.cassandra.schema.Indexes;
 import org.apache.cassandra.schema.KeyspaceParams;
 import org.apache.cassandra.utils.BloomFilter;
 import org.apache.cassandra.utils.ByteBufferUtil;
@@ -52,6 +61,8 @@ public class SensorsReadTest
     public static final String KEYSPACE1 = "SensorsReadTest";
     public static final String CF_STANDARD = "Standard";
     public static final String CF_STANDARD_CLUSTERING = "StandardClustering";
+    public static final String CF_STANDARD_SAI = "StandardSAI";
+    public static final String CF_STANDARD_SECONDARY_INDEX = "StandardSecondaryIndex";
 
     private static final String READ_BYTES_REQUEST = "READ_BYTES_REQUEST";
     private static final String READ_BYTES_TABLE = "READ_BYTES_TABLE";
@@ -62,13 +73,36 @@ public class SensorsReadTest
     @BeforeClass
     public static void defineSchema() throws Exception
     {
+        // build SAI indexes
+        Indexes.Builder saiIndexes = Indexes.builder();
+        saiIndexes.add(IndexMetadata.fromSchemaMetadata(CF_STANDARD_SAI + "_val", IndexMetadata.Kind.CUSTOM, new HashMap<>()
+        {{
+            put(IndexTarget.CUSTOM_INDEX_OPTION_NAME, StorageAttachedIndex.class.getName());
+            put(IndexTarget.TARGET_OPTION_NAME, "val");
+        }}));
+
+        // build secondary indexes
+        Indexes.Builder secondaryIndexes = Indexes.builder();
+        secondaryIndexes.add(IndexMetadata.fromIndexTargets(Collections.singletonList(
+                                                            new IndexTarget(
+                                                            new ColumnIdentifier("val", true), IndexTarget.Type.VALUES)), CF_STANDARD_SECONDARY_INDEX + "_val",
+                                                            IndexMetadata.Kind.COMPOSITES,
+                                                            Collections.emptyMap()));
+
         SchemaLoader.prepareServer();
         SchemaLoader.createKeyspace(KEYSPACE1,
                                     KeyspaceParams.simple(1),
                                     SchemaLoader.standardCFMD(KEYSPACE1, CF_STANDARD,
                                                               1, AsciiType.instance, AsciiType.instance, null),
                                     SchemaLoader.standardCFMD(KEYSPACE1, CF_STANDARD_CLUSTERING,
-                                                              1, AsciiType.instance, AsciiType.instance, AsciiType.instance));
+                                                              1, AsciiType.instance, AsciiType.instance, AsciiType.instance),
+                                    SchemaLoader.standardCFMD(KEYSPACE1, CF_STANDARD_SAI,
+                                                              1, AsciiType.instance, LongType.instance, null)
+                                                .partitioner(Murmur3Partitioner.instance) // supported by SAI
+                                                .indexes(saiIndexes.build()),
+                                    SchemaLoader.standardCFMD(KEYSPACE1, CF_STANDARD_SECONDARY_INDEX,
+                                                              1, AsciiType.instance, LongType.instance, null)
+                                                .indexes(secondaryIndexes.build()));
 
         CompactionManager.instance.disableAutoCompaction();
     }
@@ -79,13 +113,16 @@ public class SensorsReadTest
         SensorsRegistry.instance.onCreateKeyspace(Keyspace.open(KEYSPACE1).getMetadata());
         SensorsRegistry.instance.onCreateTable(Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD).metadata());
         SensorsRegistry.instance.onCreateTable(Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_CLUSTERING).metadata());
+        SensorsRegistry.instance.onCreateKeyspace(Keyspace.open(KEYSPACE1).getMetadata());
+        SensorsRegistry.instance.onCreateTable(Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_SAI).metadata());
+        SensorsRegistry.instance.onCreateTable(Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_SECONDARY_INDEX).metadata());
 
         capturedOutboundMessages = new CopyOnWriteArrayList<>();
         MessagingService.instance().outboundSink.add((message, to) ->
-        {
-            capturedOutboundMessages.add(message);
-            return false;
-        });
+                                                     {
+                                                         capturedOutboundMessages.add(message);
+                                                         return false;
+                                                     });
     }
 
     @After
@@ -93,6 +130,8 @@ public class SensorsReadTest
     {
         Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD).truncateBlocking();
         Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_CLUSTERING).truncateBlocking();
+        Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_SAI).truncateBlocking();
+        Keyspace.open(KEYSPACE1).getColumnFamilyStore(CF_STANDARD_SECONDARY_INDEX).truncateBlocking();
 
         RequestTracker.instance.set(null);
         SensorsRegistry.instance.clear();
@@ -262,6 +301,99 @@ public class SensorsReadTest
         assertThat(request2Sensor.getValue()).isEqualTo(request1Bytes * 10);
         assertThat(SensorsTestUtil.getRegistrySensor(context, Type.READ_BYTES).getValue()).isEqualTo(request1Bytes + request2Sensor.getValue());
         assertResponseSensors(request2Sensor.getValue(), request1Bytes + request2Sensor.getValue());
+    }
+
+    @Test
+    public void testSAIIndexScan()
+    {
+        store = SensorsTestUtil.discardSSTables(KEYSPACE1, CF_STANDARD_SAI);
+        Context context = new Context(KEYSPACE1, CF_STANDARD_SAI, store.metadata.id.toString());
+
+        for (int j = 0; j < 10; j++)
+        {
+            new RowUpdateBuilder(store.metadata(), j, String.valueOf(j))
+            .add("val", (long) j)
+            .build()
+            .applyUnsafe();
+        }
+
+        ReadCommand readCommand = Util.cmd(store)
+                                      .columns("val")
+                                      .filterOn("val", Operator.GT, 0L)
+                                      .build();
+
+        handleReadCommand(readCommand);
+
+        assertRequestAndRegistrySensorsEquality(context);
+
+        Sensor requestSensor = SensorsTestUtil.getThreadLocalRequestSensor(context, Type.READ_BYTES);
+        assertResponseSensors(requestSensor.getValue(), requestSensor.getValue());
+    }
+
+    @Test
+    public void testSAISingleRowSearchVSIndexScan()
+    {
+        store = SensorsTestUtil.discardSSTables(KEYSPACE1, CF_STANDARD_SAI);
+        Context context = new Context(KEYSPACE1, CF_STANDARD_SAI, store.metadata.id.toString());
+
+        int numRows = 10;
+        for (int j = 0; j < numRows; j++)
+        {
+            new RowUpdateBuilder(store.metadata(), j, String.valueOf(j))
+            .add("val", (long) j)
+            .build()
+            .applyUnsafe();
+        }
+
+        // Match a single row
+        ReadCommand readCommand = Util.cmd(store)
+                                      .columns("val")
+                                      .filterOn("val", Operator.EQ, 0L)
+                                      .build();
+        handleReadCommand(readCommand);
+
+        // Store the request sensor value for comparison with full index scan
+        Sensor indexReadSensor = SensorsTestUtil.getThreadLocalRequestSensor(context, Type.READ_BYTES);
+        double singleRowSearchBytes = indexReadSensor.getValue();
+        indexReadSensor.reset();
+
+        // Scan the whole index
+        readCommand = Util.cmd(store)
+                          .columns("val")
+                          .filterOn("val", Operator.GTE, 0L)
+                          .build();
+        handleReadCommand(readCommand);
+
+        double fullIndexScanBytes = SensorsTestUtil.getThreadLocalRequestSensor(context, Type.READ_BYTES).getValue();
+        assertThat(fullIndexScanBytes).isEqualTo(numRows * singleRowSearchBytes);
+    }
+
+    @Test
+    public void testSecondayIndexSingleRow()
+    {
+        store = SensorsTestUtil.discardSSTables(KEYSPACE1, CF_STANDARD_SECONDARY_INDEX);
+        Context context = new Context(KEYSPACE1, CF_STANDARD_SECONDARY_INDEX, store.metadata.id.toString());
+
+        for (int j = 0; j < 10; j++)
+        {
+            new RowUpdateBuilder(store.metadata(), j, String.valueOf(j))
+            .add("val", (long) j)
+            .build()
+            .applyUnsafe();
+        }
+
+        ReadCommand readCommand = Util.cmd(store)
+                                      .fromKeyIncl("0").toKeyIncl("10")
+                                      .columns("val")
+                                      .filterOn("val", Operator.EQ, 1L) // only EQ is supported by CassnadraIndex
+                                      .build();
+
+        handleReadCommand(readCommand);
+
+        assertRequestAndRegistrySensorsEquality(context);
+
+        Sensor requestSensor = SensorsTestUtil.getThreadLocalRequestSensor(context, Type.READ_BYTES);
+        assertResponseSensors(requestSensor.getValue(), requestSensor.getValue());
     }
 
     private static void handleReadCommand(ReadCommand command)
